Refactor: UserInfo에 아바타 추가, GrassArea 이동
- profile에서 GrassArea, Avatar를 관리하도록 함
- UserInfo에 profileAvatar 컴포넌트 추가
